+++ index.js
@@ -10,7 +10,7 @@ var RNNotificationsComponent = require( './component' );
 
 var RNNotifications = RNNotificationsComponent.component;
 
-var Platform = require('react-native').Platform;
+let Platform = require('react-native').Platform;
 
 var Notifications = {
   handler: RNNotifications,
@@ -437,6 +437,42 @@ Notifications.removeAllDeliveredNotifications = function() {
 
 Notifications.getDeliveredNotifications = function() {
   return this.callNative('getDeliveredNotifications', arguments);
+}
+
+Notifications.getScheduledLocalNotifications = function(callback) {
+	const mapNotifications = (notifications) => {
+		let mappedNotifications = [];
+		if(notifications?.length > 0) {
+			if(Platform.OS === 'ios'){
+				mappedNotifications = notifications.map(notif => {
+					return ({
+						soundName: notif.soundName,
+						repeatInterval: notif.repeatInterval,
+						id: notif.userInfo?.id,
+						date: new Date(notif.fireDate),
+						number: notif?.applicationIconBadgeNumber,
+						message: notif?.alertBody,
+						title: notif?.alertTitle,
+					})
+				})
+			} else if(Platform.OS === 'android') {
+				mappedNotifications = notifications.map(notif => {
+					return ({
+						soundName: notif.soundName,
+						repeatInterval: notif.repeatInterval,
+						id: notif.id,
+						date: new Date(notif.date),
+						number: notif.number,
+						message: notif.message,
+						title: notif.title,
+					})
+				})
+			}
+		}
+		callback(mappedNotifications);
+	}
+
+	return this.callNative('getScheduledLocalNotifications', [mapNotifications]);
 }
 
 Notifications.removeDeliveredNotifications = function() {